+++ src/main/java/com/braintreegateway/OAuthAccessRevocation.java
@@ -4,12 +4,18 @@ import com.braintreegateway.util.NodeWrapper;
 
 public final class OAuthAccessRevocation {
     private final String merchantId;
+    private final String oauthApplicationClientId;
 
     public OAuthAccessRevocation(NodeWrapper node) {
         this.merchantId = node.findString("merchant-id");
+        this.oauthApplicationClientId = node.findString("oauth-application-client-id");
     }
 
     public String getMerchantId() {
         return merchantId;
+    }
+
+    public String getOauthApplicationClientId() {
+        return oauthApplicationClientId;
     }
 }