+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Allow payee ID to be passed in options params for transaction create
+* Add `getMerchantId` to `ConnectedMerchantStatusTransitioned` and `ConnectedMerchantPayPalStatusChanged`, and `getOauthApplicationClientId` to OAuthAccessRevocation webhooks
 
 ## 2.80.0
 * Add support for US Bank Account verifications API